Finished IKsPropertySet
git-svn-id: https://svn.code.sf.net/p/directshownet/code/trunk@148 b2a09c84-0b36-484d-b36e-ffab392acda3

+++ directshowlib/src/DsUtils.cs
@@ -849,7 +849,7 @@ namespace DirectShowLib
                 if (pKs != null)
                 {
                     // Query for the Category
-                    hr = pKs.Get(g, AMPropertyPin.Category, IntPtr.Zero, 0, ipOut, iSize, out cbBytes);
+                    hr = pKs.Get(g, (int)AMPropertyPin.Category, IntPtr.Zero, 0, ipOut, iSize, out cbBytes);
                     DsError.ThrowExceptionForHR(hr);
 
                     // Marshal it to the return variable
